# Optimized v1 to v2 DB upgrade