Huge thanks to @neurosis69 for helping me test on his setups, especially the RPi4 one which allowed me to address the fact that our existing upgrade tool never finishes.