@arvidn 2024-03-29
Comparing this to `main`.

| version | timing | fraction |
|--- | --- | --- |
| main | 4:32:01 (16320s) | 100% |
| This PR | 2:49:46 (10186s) | 62.4% |

```
there is very little free space on the volume where the output database will be written:
   /home/arvid/blockchain_v2_mainnet.sqlite
free space: 305.64 GiB expected about 309.59 GiB
opening file for reading: /home/arvid/blockchain_v1_mainnet.sqlite
opening file for writing: /home/arvid/blockchain_v2_mainnet.sqlite
initializing v2 version
initializing v2 block store
peak: 85afa3e553d069236d2807f419641a1720777d2a6fd6279039ab44e5ad545b94 height: 5040549
[1/5] converting full_blocks
      1618.47 seconds
[2/5] converting sub_epoch_segments_v3
      51.57 seconds
[3/5] converting hint_store
      596.79 seconds
[4/5] converting coin_store
      12020.08 seconds                             
[5/5] build indices
      block store
      coin store
      hint store
      2033.23 seconds                             


LEAVING PREVIOUS DB FILE UNTOUCHED /home/arvid/blockchain_v1_mainnet.sqlite

        Command being timed: "chia db upgrade --input /home/arvid/blockchain_v1_mainnet.sqlite --output /home/arvid/blockchain_v2_mainnet.sqlite"
        User time (seconds): 6498.67
        System time (seconds): 8077.26
        Percent of CPU this job got: 89%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 4:32:01
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 2239048
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 3004808
        Voluntary context switches: 19096210
        Involuntary context switches: 79769
        Swaps: 0
        File system inputs: 1039312368
        File system outputs: 2842889992
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```